chatbot-scada/requirements.txt
absl-py==0.13.0
aio-pika==6.8.2
aiofiles==23.1.0
aiohttp==3.7.4
aiormq==3.3.1
APScheduler==3.7.0
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
backcall==0.2.0
bidict==0.22.1
boto3==1.28.19
botocore==1.31.19
CacheControl==0.12.14
cachetools==4.2.4
certifi==2023.7.22
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.6
cloudpickle==1.6.0
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.0.4
comm==0.1.4
cryptography==41.0.3
cycler==0.11.0
dask==2021.11.2
debugpy==1.6.7
decorator==5.1.1
dnspython==1.16.0
docopt==0.6.2
executing==1.2.0
fbmessenger==6.0.0
fire==0.5.0
flatbuffers==2.0.7
fsspec==2023.6.0
future==0.18.3
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.56.2
h5py==3.9.0
httptools==0.6.0
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.1
ipython==8.12.2
jedi==0.19.0
jmespath==1.0.1
joblib==1.0.1
jsonpickle==2.0.0
jsonschema==3.2.0
jupyter-client==8.3.0
jupyter-core==5.3.1
kafka-python==2.0.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==16.0.6
locket==1.0.0
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.3.4
matplotlib-inline==0.1.6
mattermostwrapper==2.2
msgpack==1.0.5
multidict==5.2.0
nest-asyncio==1.5.7
networkx==2.6.3
nltk==3.8.1
numpy==1.19.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==20.9
pamqp==2.3.0
parso==0.8.3
partd==1.4.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
protobuf==3.19.6
psutil==5.9.5
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydot==1.4.2
Pygments==2.16.1
PyJWT==2.8.0
pykwalify==1.8.0
pymongo==3.10.1
pyparsing==3.1.1
pyrsistent==0.19.3
pyTelegramBotAPI==3.8.3
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-engineio==4.5.1
python-socketio==5.8.0
pytz==2021.3
PyYAML==6.0.1
pyzmq==25.1.0
questionary==1.10.0
randomname==0.1.5
rasa==3.1.0
rasa-sdk==3.1.3
redis==3.5.3
regex==2021.8.28
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rocketchat-API==1.16.0
rsa==4.9
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.7
s3transfer==0.6.1
sanic==21.12.2
Sanic-Cors==2.2.0
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==0.24.2
scipy==1.10.1
sentry-sdk==1.3.1
six==1.16.0
sklearn-crfsuite==0.3.6
slackclient==2.9.4
SQLAlchemy==1.4.49
stack-data==0.6.2
tabulate==0.9.0
tarsafe==0.0.3
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.7.4
tensorflow-addons==0.15.0
tensorflow-estimator==2.7.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.33.0
tensorflow-text==2.13.0
termcolor==2.3.0
terminaltables==3.1.10
threadpoolctl==3.2.0
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
twilio==6.50.1
typeguard==4.1.0
typing-extensions==4.7.1
typing-utils==0.1.0
tzlocal==2.1
ujson==4.3.0
underthesea==6.7.0
underthesea-core==1.0.4
urllib3==2.0.4
uvloop==0.17.0
wcwidth==0.2.6
webexteamssdk==1.6.1
websockets==10.0
Werkzeug==2.3.6
wrapt==1.15.0
yarl==1.9.2
zipp==3.16.2
